+++ src/Oro/Bundle/ApiBundle/Tests/Unit/Processor/CollectSubresources/LoadFromConfigBagTest.php
@@ -2,12 +2,12 @@
 
 namespace Oro\Bundle\ApiBundle\Tests\Unit\Processor\CollectSubresources;
 
-use Oro\Bundle\ApiBundle\Config\ActionsConfigLoader;
 use Oro\Bundle\ApiBundle\Config\Config;
-use Oro\Bundle\ApiBundle\Config\ConfigLoaderFactory;
 use Oro\Bundle\ApiBundle\Config\EntityDefinitionConfig;
-use Oro\Bundle\ApiBundle\Config\EntityDefinitionConfigExtra;
-use Oro\Bundle\ApiBundle\Config\SubresourcesConfigLoader;
+use Oro\Bundle\ApiBundle\Config\Extra\EntityDefinitionConfigExtra;
+use Oro\Bundle\ApiBundle\Config\Loader\ActionsConfigLoader;
+use Oro\Bundle\ApiBundle\Config\Loader\ConfigLoaderFactory;
+use Oro\Bundle\ApiBundle\Config\Loader\SubresourcesConfigLoader;
 use Oro\Bundle\ApiBundle\Metadata\AssociationMetadata;
 use Oro\Bundle\ApiBundle\Metadata\EntityMetadata;
 use Oro\Bundle\ApiBundle\Model\EntityIdentifier;