+++ src/Oro/Bundle/ApiBundle/Tests/Unit/Processor/TestConfigExtra.php
@@ -2,7 +2,7 @@
 
 namespace Oro\Bundle\ApiBundle\Tests\Unit\Processor;
 
-use Oro\Bundle\ApiBundle\Config\ConfigExtraInterface;
+use Oro\Bundle\ApiBundle\Config\Extra\ConfigExtraInterface;
 use Oro\Bundle\ApiBundle\Processor\Config\ConfigContext;
 
 class TestConfigExtra implements ConfigExtraInterface